+++ .env.example
@@ -2,4 +2,5 @@ DB_HOST=
 DB_USER=
 DB_PASSWORD=
 DB_NAME=
-SECRET_KEY=
+ACCESS_SECRET_KEY=
+REFRESH_SECRET_KEY=
