+++ helpers/index.js
@@ -1,0 +1,4 @@
+const generateAccessToken = require("./generateAccessToken");
+const generateRefreshToken = require("./generateRefreshToken");
+
+module.exports = { generateAccessToken, generateRefreshToken };
